Force certain tasks to run synchronously
A new setting was added to the @background decorator that forces that
specific task to run synchronously even if BACKGROUND_TASK_RUN_ASYNC is
True. This ensures that at any time, only one instance of this task is
running. Tasks with this setting that are enqueued when another task of
its type is executing are backlogged and will be attempted to run later
once the existing task completes.

Usage:

    from background_task.tasks import background

    @background(name='aName', force_synchronous_execution=True)
    def your_task():
        pass